Move transformation into operations/
